remove intermediary wire for taps
