+++ src/test/scala/chiselTests/BoringUtilsSpec.scala
@@ -3,7 +3,6 @@
 package chiselTests
 
 import chisel3._
-import chisel3.probe
 import chisel3.util.Counter
 import chisel3.testers._
 import chisel3.experimental.{BaseModule, ChiselAnnotation}
@@ -325,222 +324,5 @@ class BoringUtilsSpec extends ChiselFlatSpec with ChiselRunners with Utils with 
     matchesAndOmits(chirrtl)()(
       "connect a_bore, a"
     )
-  }
-
-  "Downwards tap from parent to child" should "work" in {
-    class Foo extends RawModule {
-      val internalWire = Wire(Bool())
-    }
-    class Top extends RawModule {
-      val foo = Module(new Foo())
-      val outProbe = IO(probe.Probe(Bool()))
-      val out = IO(Bool())
-      probe.define(outProbe, BoringUtils.tap(foo.internalWire))
-      out := BoringUtils.tapAndRead(foo.internalWire)
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Foo :",
-      "output bore : Probe<UInt<1>>",
-      "output out_bore : Probe<UInt<1>>",
-      "define bore = probe(internalWire)",
-      "define out_bore = probe(internalWire)",
-      "module Top :",
-      "define outProbe = foo.bore",
-      "connect out, read(foo.out_bore)"
-    )()
-  }
-
-  "Downwards tap from grandparent to grandchild" should "work" in {
-    class Bar extends RawModule {
-      val internalWire = Wire(Bool())
-    }
-    class Foo extends RawModule {
-      val bar = Module(new Bar)
-    }
-    class Top extends RawModule {
-      val foo = Module(new Foo)
-      val out = IO(Bool())
-      out := BoringUtils.tapAndRead(foo.bar.internalWire)
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Bar :",
-      "output out_bore : Probe<UInt<1>>",
-      "define out_bore = probe(internalWire)",
-      "module Foo :",
-      "output out_bore : Probe<UInt<1>>",
-      "define out_bore = bar.out_bore",
-      "module Top :",
-      "connect out, read(foo.out_bore)"
-    )()
-  }
-
-  "Upwards tap from child to parent" should "work" in {
-    class Foo(parentData: Data) extends RawModule {
-      val outProbe = IO(probe.Probe(Bool()))
-      val out = IO(Bool())
-      probe.define(outProbe, BoringUtils.tap(parentData))
-      out := BoringUtils.tapAndRead(parentData)
-    }
-    class Top extends RawModule {
-      val parentWire = Wire(Bool())
-      val foo = Module(new Foo(parentWire))
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Foo :",
-      "input bore : UInt<1>",
-      "input out_bore : UInt<1>",
-      "define outProbe = probe(tapIntermediate)",
-      "connect out, out_tapIntermediate",
-      "module Top :",
-      "connect foo.bore, parentWire",
-      "connect foo.out_bore, parentWire"
-    )()
-  }
-
-  "Upwards tap from grandchild to grandparent" should "work" in {
-    class Bar(grandParentData: Data) extends RawModule {
-      val out = IO(Bool())
-      out := BoringUtils.tapAndRead(grandParentData)
-    }
-    class Foo(parentData: Data) extends RawModule {
-      val bar = Module(new Bar(parentData))
-    }
-    class Top extends RawModule {
-      val parentWire = Wire(Bool())
-      val foo = Module(new Foo(parentWire))
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Bar :",
-      "input out_bore : UInt<1>",
-      "connect out, out_tapIntermediate",
-      "module Foo :",
-      "connect bar.out_bore, out_bore",
-      "input out_bore : UInt<1>",
-      "module Top :",
-      "connect foo.out_bore, parentWire"
-    )()
-  }
-
-  "Tap from child to sibling" should "work" in {
-    class Bar extends RawModule {
-      val a = Wire(Bool())
-    }
-    class Baz(_a: Bool) extends RawModule {
-      val b = Wire(Bool())
-      b := BoringUtils.tapAndRead(_a)
-    }
-    class Top extends RawModule {
-      val bar = Module(new Bar)
-      val baz = Module(new Baz(bar.a))
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Bar :",
-      "output b_bore : Probe<UInt<1>>",
-      "define b_bore = probe(a)",
-      "module Baz :",
-      "input b_bore : UInt<1>",
-      "connect b_tapIntermediate, b_bore",
-      "module Top :",
-      "connect baz.b_bore, read(bar.b_bore)"
-    )()
-  }
-
-  "Tap from child to sibling at different levels" should "work" in {
-    class Bar extends RawModule {
-      val a = Wire(Bool())
-    }
-    class Baz(_a: Bool) extends RawModule {
-      val b = Wire(Bool())
-      b := BoringUtils.tapAndRead(_a)
-    }
-    class Foo(_a: Bool) extends RawModule {
-      val baz = Module(new Baz(_a))
-    }
-    class Top extends RawModule {
-      val bar = Module(new Bar)
-      val foo = Module(new Foo(bar.a))
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Bar :",
-      "output b_bore : Probe<UInt<1>>",
-      "define b_bore = probe(a)",
-      "module Baz :",
-      "input b_bore : UInt<1>",
-      "connect b_tapIntermediate, b_bore",
-      "module Foo :",
-      "input b_bore : UInt<1>",
-      "connect baz.b_bore, b_bore",
-      "module Top :",
-      "connect foo.b_bore, read(bar.b_bore)"
-    )()
-  }
-
-  "Downwards writable tap from grandparent to grandchild" should "work" in {
-    class Bar extends RawModule {
-      val internalWire = Wire(Bool())
-    }
-    class Foo extends RawModule {
-      val bar = Module(new Bar)
-    }
-    class Top extends RawModule {
-      val foo = Module(new Foo)
-      val out = IO(Bool())
-      out := probe.read(BoringUtils.rwTap(foo.bar.internalWire))
-      probe.forceInitial(BoringUtils.rwTap(foo.bar.internalWire), false.B)
-    }
-    val chirrtl = circt.stage.ChiselStage.emitCHIRRTL(new Top)
-    matchesAndOmits(chirrtl)(
-      "module Bar :",
-      "output out_bore : RWProbe<UInt<1>>",
-      "define out_bore = rwprobe(internalWire)",
-      "module Foo :",
-      "output out_bore : RWProbe<UInt<1>>",
-      "define out_bore = bar.out_bore",
-      "module Top :",
-      "connect out, read(foo.out_bore)",
-      "force_initial(foo.bore, UInt<1>(0h0))"
-    )()
-  }
-
-  "Upwards writable tap from child to parent" should "not work" in {
-    class Foo(parentData: Data) extends RawModule {
-      val outProbe = IO(probe.RWProbe(Bool()))
-      probe.define(outProbe, BoringUtils.rwTap(parentData))
-    }
-    class Top extends RawModule {
-      val parentWire = Wire(Bool())
-      val foo = Module(new Foo(parentWire))
-    }
-    val e = intercept[Exception] {
-      circt.stage.ChiselStage.emitCHIRRTL(new Top, Array("--throw-on-first-error"))
-    }
-    e.getMessage should include("Cannot drill writable probes upwards.")
-  }
-
-  "Writable tap from child to sibling at different levels" should "not work" in {
-    class Bar extends RawModule {
-      val a = Wire(Bool())
-    }
-    class Baz(_a: Bool) extends RawModule {
-      val b = Output(probe.RWProbe(Bool()))
-      b := BoringUtils.rwTap(_a)
-    }
-    class Foo(_a: Bool) extends RawModule {
-      val baz = Module(new Baz(_a))
-    }
-    class Top extends RawModule {
-      val bar = Module(new Bar)
-      val foo = Module(new Foo(bar.a))
-    }
-    val e = intercept[Exception] {
-      circt.stage.ChiselStage.emitCHIRRTL(new Top, Array("--throw-on-first-error"))
-    }
-    e.getMessage should include("Cannot drill writable probes upwards.")
   }
 }
